Refactor data types for IDs to Long in Collection and CollectionComment models; update related API methods and navigation arguments

+++ app/src/main/java/com/example/finalproject/adapters/CommentAdapter.kt
@@ -9,13 +9,11 @@ import com.bumptech.glide.Glide
 import com.example.finalproject.R
 import com.example.finalproject.databinding.ItemCommentBinding
 import com.example.finalproject.models.CollectionComment
-// import java.text.SimpleDateFormat // Для форматирования даты
-// import java.util.Locale
 
-class CommentAdapter(private val onAuthorClicked: (Int?) -> Unit) :
+class CommentAdapter(private val onAuthorClicked: (Long?) -> Unit) :
     ListAdapter<CollectionComment, CommentAdapter.CommentViewHolder>(CommentDiffCallback()) {
 
-    // private val dateFormat = SimpleDateFormat("dd MMM yyyy, HH:mm", Locale.getDefault()) // Пример
+    // private val dateFormat = SimpleDateFormat("dd MMM yyyy, HH:mm", Locale.getDefault())
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CommentViewHolder {
         val binding = ItemCommentBinding.inflate(LayoutInflater.from(parent.context), parent, false)
@@ -28,14 +26,14 @@ class CommentAdapter(private val onAuthorClicked: (Int?) -> Unit) :
 
     class CommentViewHolder(
         private val binding: ItemCommentBinding,
-        private val onAuthorClicked: (Int?) -> Unit
+        private val onAuthorClicked: (Long?) -> Unit
     ) : RecyclerView.ViewHolder(binding.root) {
         fun bind(comment: CollectionComment) {
             binding.commentText.text = comment.commentText
             binding.commentAuthorName.text = comment.user?.displayName ?: comment.user?.username ?: itemView.context.getString(R.string.unknown_author)
             // binding.commentDate.text = comment.createdAt // TODO: Format date properly
             // Простая заглушка для даты
-            binding.commentDate.text = " " // Замените на реальное форматирование даты
+            binding.commentDate.text = " "
 
             Glide.with(itemView.context)
                 .load(comment.user?.avatarUrl)
