+++ app/src/main/java/com/example/finalproject/models/CollectionComment.kt
@@ -2,8 +2,8 @@ package com.example.finalproject.models
 
 data class CollectionComment(
     val commentId: Int,
-    val collectionId: Int,
-    val userId: Int?,
+    val collectionId: Long,
+    val userId: Long?,
     val commentText: String,
     val createdAt: String, // Или использовать тип Date/Timestamp
     val user: User? = null // Детали пользователя, оставившего комментарий
